feat: add to_rfc1123_datetime helper function; create datetime_utils
