+++ eyes_common/applitools/common/utils/__init__.py
@@ -11,6 +11,11 @@ from .compat import (  # type: ignore # noqa
     urlsplit,
     urlunsplit,
 )
+from .datetime_utils import (  # type: ignore # noqa
+    UTC,
+    current_time_in_rfc1123,
+    to_rfc1123_datetime,
+)
 from .general_utils import cached_property  # noqa
 
 __all__ = compat.__all__ + ("image_utils", "argument_guard")  # noqa
